Updated the formatting. Helps to save the file after formatting it in Eclipse.
